Fix a signedness problem in zstty_cncheckc(): when no character is ready, -1 should be returned, but it was assigned to an uint8_t (which is extended to an int to form the return value), causing 255 to be returned instead.

+++ sys/dev/zs/zs.c
@@ -943,7 +943,7 @@ zstty_cngetc(struct zstty_softc *sc)
 static int
 zstty_cncheckc(struct zstty_softc *sc)
 {
-	uint8_t c;
+	int c;
 
 	c = -1;
 	zstty_cnopen(sc);
